tests/requirements.txt
pytest-playwright>=0.7.0,<0.8
pytest>=8.3.5,<9.0
python-dotenv>=1.1.0,<2.0
